requirements.txt
tensorflow
boto3
Pillow==9.5.0